+++ src/main/java/com/simpleschedule/daniel/anderson/services/ContactService.java
@@ -18,22 +18,51 @@ public class ContactService {
 	
 	//SAVE NEW CONTACT
 	public Contact saveNewContact(Contact contact) {
-		return contactRepository.save(contact);
+		if (contact != null) {
+			return contactRepository.save(contact);
+		} else {
+			return null;
+		}
 	}
 	
 	//FIND CONTACT BY cPatientId
 	public Contact findContactBycPatientId(Integer cPatientId) {
-		return contactRepository.findByCPatientId(cPatientId);
+		if (cPatientId != null) {
+			Contact foundContact = contactRepository.findByCPatientId(cPatientId);
+			if (foundContact != null) {
+				return foundContact;
+			} else {
+				return null;
+			}
+		} else {
+			return null;
+		}
 	}
 	
 	//UPDATE CONTACT
 	public Contact updateContact(Contact updateContact) {
-		return contactRepository.save(updateContact);
+		if (updateContact != null) {
+			if (contactRepository.findByCPatientId(updateContact.getcPatientId()) != null) {
+				return contactRepository.save(updateContact);
+			} else {
+				return null;
+			}
+		} else {
+			return null;
+		}
 	}
 	
 	//DELETE CONTACT
-	public void deleteContact(Contact deleteContact) {
-		contactRepository.delete(deleteContact);
-		System.out.println("REMOVED: " + deleteContact);
+	public boolean deleteContact(Contact deleteContact) {
+		if (deleteContact != null) {
+			if (contactRepository.findByCPatientId(deleteContact.getcPatientId()) != null) {
+				contactRepository.delete(deleteContact);
+				return true;
+			} else {
+				return false;
+			}
+		} else {
+			return false;
+		}
 	}
 }
